Fix all leaderboard duplicated
It only happen when there is no internet connection. When there is no internet connection, data in the leaderboard cannot be loaded from backend, so it will take whatever inside the cache and display on the leaderboard.
My solution for this is to clear the leaderboard when dispatch. Every time before loading the leaderboard, the leaderboard list will be clear. Hence, when there is no internet connection, instead of displaying the same leaderboard in the cache, it will display a blank leaderboard.